Fixed declaration problem
declaration problem in exception.h

+++ Kernel/include/exceptions.h
@@ -1,7 +1,7 @@
 #ifndef EXCEPTIONS_H_
 #define EXCEPTIONS_H_
 
-void exceptionDispatcher(int exception, uint64_t* rsp);
+void exceptionDispatcher(uint64_t exception, uint64_t* rsp);
 void zero_division(uint64_t* rsp);
 void overflow(uint64_t* rsp);
 void opcode(uint64_t* rsp);
